# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MicroLens Recommendation MVP API",
    "description": "Phase 2D public runtime. Feed orchestration remains explicitly deferred to Phase 4.",
    "version": "0.2.0-phase2d",
    "x-implementation-status": "phase_2d_public_runtime"
  },
  "paths": {
...
    "/api/admin/dashboard/export.csv": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Export Csv",
        "operationId": "exportDashboardCsv",
        "parameters": [
          {
            "name": "from_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From Utc"
            }
          },
          {
            "name": "to_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To Utc"
            }
          },
          {
            "name": "feed_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FeedType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
...
    "/api/feeds/{feed_type}": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "Deferred Feed",
        "operationId": "getFeedPage",
        "parameters": [
          {
            "name": "feed_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FeedType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPage"
                }
              }
            }
          },
          "501": {
            "description": "Deferred to Phase 4",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "user",
          "operator_readonly",
          "operator",
          "admin"
        ],
        "x-implementation-phase": "phase_4_deferred"
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ErrorEnvelope"
      },
...
      "FeedItem": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "title",
          "cover",
          "position",
          "source",
          "score",
          "reason",
          "model_version"
        ],
        "title": "FeedItem"
      },
...
      "FeedPage": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot_id",
          "request_id",
          "model_version",
          "items",
          "next_cursor"
        ],
        "title": "FeedPage"
      },
      "FeedType": {
        "type": "string",
        "enum": [
          "personalized",
          "popular",
          "explore"
        ],
        "title": "FeedType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "microlens_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "Error": {
        "description": "Structured API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Host-published public listener"
    }
  ],
...
}